# server/openapi.yaml
openapi: 3.1.0
info:
  title: WealthGenie API
  version: 1.0.0
  description: >
    Production-grade financial planning and investment advice engine API.
    Provides automated portfolio recommendation, Monte Carlo wealth projection,
    Indian tax optimization (FY2025-26), and XIRR goal calculations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local Development Server
  - url: https://api.wealthgenie.io
    description: Production Cluster Gateway
paths:
...
  /api/goals:
    get:
      operationId: listGoals
      summary: List User Goals
      description: Returns all financial goals for the authenticated user.
      tags:
        - Goals
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Goals list retrieved
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createGoal
      summary: Create Financial Goal
      description: Creates a new financial goal with reverse SIP calculation.
      tags:
        - Goals
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalRequest'
      responses:
        '201':
          description: Goal created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/goals/{id}:
    put:
      operationId: updateGoal
      summary: Update Goal
      tags:
        - Goals
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Goal updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      operationId: deleteGoal
      summary: Delete Goal
      tags:
        - Goals
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Goal deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request / validation error
      content:
        application/json:
          example:
            error: Invalid request parameters
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          example:
            error: Access denied. No token provided.
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            error: Requested resource not found
...
  schemas:
...
    GoalRequest:
      type: object
      required: [goal_name, target_amount, horizon_years]
      properties:
        goal_name:
          type: string
          example: Retirement Fund
        target_amount:
          type: number
          minimum: 1000
          example: 10000000
        horizon_years:
          type: integer
          minimum: 1
          maximum: 40
          example: 15